Modified api file structure/routes
