+++ client/components/CardsetView.js
@@ -17,7 +17,7 @@ export const CardsetView = ({cardset}) => {
 
     const fetchFlashCards = async () => {
         try{
-            const response = await axios.get(process.env.NEXT_PUBLIC_SERVER_URL+'/api/getflashcards',  {params: { cardsetId: cardset.id}});
+            const response = await axios.get(process.env.NEXT_PUBLIC_SERVER_URL+`/api/flashcards/${cardset.id}`);
             const flashcards = response.data.flashcards;
             setCurrentCardsetData(flashcards);
         } catch(error) {
